+++ x/oracle/abci.go
@@ -1,6 +1,7 @@
 package oracle
 
 import (
+	"strings"
 	"time"
 
 	"github.com/cosmos/cosmos-sdk/telemetry"
@@ -10,17 +11,12 @@ import (
 	"github.com/umee-network/umee/v4/x/oracle/types"
 )
 
-// isPeriodLastBlock returns true if we are at the last block of the period
-func isPeriodLastBlock(ctx sdk.Context, blocksPerPeriod uint64) bool {
-	return (uint64(ctx.BlockHeight())+1)%blocksPerPeriod == 0
-}
-
 // EndBlocker is called at the end of every block
 func EndBlocker(ctx sdk.Context, k keeper.Keeper) error {
 	defer telemetry.ModuleMeasureSince(types.ModuleName, time.Now(), telemetry.MetricKeyEndBlocker)
 
 	params := k.GetParams(ctx)
-	if isPeriodLastBlock(ctx, params.VotePeriod) {
+	if k.IsPeriodLastBlock(ctx, params.VotePeriod) {
 		// Build claim map over all validators in active set
 		validatorClaimMap := make(map[string]types.Claim)
 		powerReduction := k.StakingKeeper.PowerReduction(ctx)
@@ -46,6 +42,7 @@ func EndBlocker(ctx sdk.Context, k keeper.Keeper) error {
 
 		// Iterate through ballots and update exchange rates; drop if not enough votes have been achieved.
 		for _, ballotDenom := range ballotDenomSlice {
+			denom := strings.ToUpper(ballotDenom.Denom)
 			// Get weighted median of exchange rates
 			exchangeRate, err := Tally(ballotDenom.Ballot, params.RewardBand, validatorClaimMap)
 			if err != nil {
@@ -53,17 +50,17 @@ func EndBlocker(ctx sdk.Context, k keeper.Keeper) error {
 			}
 
 			// Set the exchange rate, emit ABCI event
-			if err = k.SetExchangeRateWithEvent(ctx, ballotDenom.Denom, exchangeRate); err != nil {
+			if err = k.SetExchangeRateWithEvent(ctx, denom, exchangeRate); err != nil {
 				return err
 			}
 
-			if isPeriodLastBlock(ctx, params.HistoricStampPeriod) {
-				k.AddHistoricPrice(ctx, ballotDenom.Denom, exchangeRate)
+			if k.IsPeriodLastBlock(ctx, params.HistoricStampPeriod) {
+				k.AddHistoricPrice(ctx, denom, exchangeRate)
 			}
 
 			// Calculate and stamp median/median deviation if median stamp period has passed
-			if isPeriodLastBlock(ctx, params.MedianStampPeriod) {
-				if err = k.CalcAndSetHistoricMedian(ctx, ballotDenom.Denom); err != nil {
+			if k.IsPeriodLastBlock(ctx, params.MedianStampPeriod) {
+				if err = k.CalcAndSetHistoricMedian(ctx, denom); err != nil {
 					return err
 				}
 			}
@@ -98,21 +95,11 @@ func EndBlocker(ctx sdk.Context, k keeper.Keeper) error {
 
 	// Slash oracle providers who missed voting over the threshold and
 	// reset miss counters of all validators at the last block of slash window
-	if isPeriodLastBlock(ctx, params.SlashWindow) {
+	if k.IsPeriodLastBlock(ctx, params.SlashWindow) {
 		k.SlashAndResetMissCounters(ctx)
 	}
 
-	// Prune historic prices and medians outside pruning period determined by
-	// the stamp period multiplied by the max stamps.
-	if isPeriodLastBlock(ctx, params.HistoricStampPeriod) {
-		pruneHistoricPeriod := params.HistoricStampPeriod*(params.MaximumPriceStamps) - params.VotePeriod
-		pruneMedianPeriod := params.MedianStampPeriod*(params.MaximumMedianStamps) - params.VotePeriod
-		for _, v := range params.AcceptList {
-			k.DeleteHistoricPrice(ctx, v.SymbolDenom, uint64(ctx.BlockHeight())-pruneHistoricPeriod)
-			k.DeleteHistoricMedian(ctx, v.SymbolDenom, uint64(ctx.BlockHeight())-pruneMedianPeriod)
-			k.DeleteHistoricMedianDeviation(ctx, v.SymbolDenom, uint64(ctx.BlockHeight())-pruneMedianPeriod)
-		}
-	}
+	k.PruneAllPrices(ctx)
 
 	return nil
 }