+++ x/oracle/types/price.go
@@ -1,6 +1,8 @@
 package types
 
 import (
+	"sort"
+
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
 
@@ -22,4 +24,38 @@ func (p *Prices) Decs() []sdk.Dec {
 		decs = append(decs, price.ExchangeRateTuple.ExchangeRate)
 	}
 	return decs
+}
+
+func (p *Prices) FilterByBlock(blockNum uint64) *Prices {
+	prices := Prices{}
+	for _, price := range *p {
+		if price.BlockNum == blockNum {
+			prices = append(prices, price)
+		}
+	}
+	return &prices
+}
+
+func (p *Prices) FilterByDenom(denom string) *Prices {
+	prices := Prices{}
+	for _, price := range *p {
+		if price.ExchangeRateTuple.Denom == denom {
+			prices = append(prices, price)
+		}
+	}
+	return &prices
+}
+
+func (p *Prices) Sort() *Prices {
+	prices := *p
+	sort.Slice(
+		prices,
+		func(i, j int) bool {
+			if prices[i].BlockNum == prices[j].BlockNum {
+				return prices[i].ExchangeRateTuple.Denom < prices[j].ExchangeRateTuple.Denom
+			}
+			return prices[i].BlockNum < prices[j].BlockNum
+		},
+	)
+	return &prices
 }